+++ lib/distribute_reads/appropriate_pool.rb
@@ -2,7 +2,8 @@ module DistributeReads
   module AppropriatePool
     def _appropriate_pool(*args)
       if Thread.current[:distribute_reads]
-        if needs_master?(*args) || @slave_pool.completely_blacklisted?
+        if needs_master?(*args) || (blacklisted = @slave_pool.completely_blacklisted?)
+          raise DistributeReads::NoReplicasAvailable, "No replicas available" if blacklisted && Thread.current[:distribute_reads][:failover] == false
           stick_to_master(*args) unless DistributeReads.default_to_primary
           @master_pool
         elsif in_transaction?