+++ lib/distribute_reads/global_methods.rb
@@ -1,6 +1,6 @@
 module DistributeReads
   module GlobalMethods
-    def distribute_reads(max_lag: nil)
+    def distribute_reads(max_lag: nil, failover: true)
       raise ArgumentError, "Missing block" unless block_given?
 
       if max_lag && DistributeReads.lag > max_lag
@@ -9,7 +9,7 @@ module DistributeReads
 
       previous_value = Thread.current[:distribute_reads]
       begin
-        Thread.current[:distribute_reads] = true
+        Thread.current[:distribute_reads] = {failover: failover}
         value = yield
         warn "[distribute_reads] Call `to_a` inside block to execute query on replica" if value.is_a?(ActiveRecord::Relation) && !previous_value
         value